+++ rose-network-irose/src/world_client_packets.rs
@@ -42,7 +42,7 @@ impl<'a> TryFrom<&'a Packet> for PacketClientConnectRequest<'a> {
 impl<'a> From<&'a PacketClientConnectRequest<'a>> for Packet {
     fn from(packet: &'a PacketClientConnectRequest<'a>) -> Self {
         let mut writer = PacketWriter::new(ClientPackets::ConnectRequest as u16);
-        writer.write_u32(packet.login_token as u32);
+        writer.write_u32(packet.login_token);
         writer.write_fixed_length_utf8(packet.password_md5, 32);
         writer.into()
     }
