+++ src/game/systems/command_system.rs
@@ -758,8 +758,7 @@ pub fn command_system(
 
                         let in_distance = target_position
                             .map(|target_position| {
-                                position.position.xy().distance(target_position.xy())
-                                    < cast_range as f32
+                                position.position.xy().distance(target_position.xy()) < cast_range
                             })
                             .unwrap_or(true);
                         if in_distance {
